refactor: InternalArrow/ArrowPathResult を src/lib/types.ts に移動し lib→features のレイヤー依存を解消
